+++ CHANGELOG.md
@@ -1,5 +1,42 @@
-# 2.28.1-0 / Unreleased
-* [Fixed] #848 Drag End Error with dnd5 extension
+# 2.30.0-0 / Unreleased
+  * [Changed] ext-edit trigger 'clickActive' now only triggers if no modifier keys
+    (shift, meta, control, ...) are pressed.<br>
+    Trigger 'shift+click' now only triggers if no other modifier key (control, ...)
+    is pressed.
+  * [Added] ext-multi is now deployed with jquery.fancytree-all.js (still experimental)
+  * [Added] tree.activateKey(key, opts) now has an `opts` argument
+  * [Added] `nodata` option (bool, string, or callback)
+  * [Added] ext-table `mergeStatusColumns` option
+  * [Fixed] ES6 import dependency on jquery for jquery.fancytree.ui-deps.js
+  * [Fixed] #863 setActive() sometimes does not scroll node into view
+  * [Fixed] #877 postProcess may now also return the object form `{..., children: []}`
+    (before, only an array `[...]` was handled correctly).
+  * [Fixed] autoScroll, node.scrollIntoView(), and .makeVisible() now work for tables as well.
+  * [Changed] #879 Rename ext-debug to ext-logger
+    (jquery.fancytree.debug.js => jquery.fancytree.logger.js)
+
+# 2.29.1 / 2018-06-27
+  * [Fixed] #848 Drag End Error with dnd5 extension (again):
+    fancytree-drag-remove class not removed on drop/dragend
+  * [Fixed] #875 ext-dnd5: Unwanted expanding of folder node when a node is dragged
+    before/after it
+  * [Fixed] #876 `triggerStart: []` does not override the default settings.<br>
+    **NOTE:** Options of type `Array` will now override the default option.
+    Before, arrays were merged with the default.
+  * [Fixed] ext-ariagrid default actions
+
+# 2.29.0 / 2018-06-16
+  * [Changed]
+    `toggleEffect` now also accepts "toggle" or "slideToggle" to use jQuery effects instead of jQueryUI.<br>
+    `toggleEffect: { effect: "slideToggle", duration: 200 }` is now the default.<br>
+    'effects' component was removed from the bundled jquery.fancytree.ui-deps.js
+  * [Fixed] #746 Animation bug when expanding/collapsing nodes
+  * [Fixed] #848 Drag End Error with dnd5 extension
+  * [Fixed] #850 ext-childcounter doesn't work with custom icons
+  * [Fixed] #859 Fix log level configuration problem
+  * [Fixed] #865 toggleEffect animation (effect: blind) sometimes got stuck.
+  * Stop testing jQuery UI 1.9
+  * Update to jQuery 3.3.1
 
 # 2.28.1 / 2018-03-19
   * [Fixed] #844 Fix RTL for ext-table